+++ activerecord/lib/active_record/base.rb
@@ -322,6 +322,7 @@ module ActiveRecord #:nodoc:
     include Store
     include SecureToken
     include Suppressor
+    include DestroyLater
   end
 
   ActiveSupport.run_load_hooks(:active_record, Base)